Do not use backticks for table names in between scope
This is a reimplementation of PR #406 by @jurgens that has been rebased
with master and with a passing build.

The use of backticks for table names is specific to the MySQL syntax.
Other DBMSes like PostgreSQL use double quotes. However, all DBMSes
should get along fine with only the table name in the query.